# RestApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Superhero API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/creategadget": {
            "post": {
                "tags": [
                    "Gadget"
                ],
                "summary": "Create a gadget",
                "operationId": "d1e874091b0557eb3200a993b37b6508",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the gadget",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "example of the body request",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bow"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gadget created"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "409": {
                        "description": "Gadget already exists"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        },
        "/api/gadget/{id}": {
            "get": {
                "tags": [
                    "Gadget"
                ],
                "summary": "Get a gadget by ID",
                "operationId": "0eb13b1abc3963c2c8f6f1947f49bf3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of gadget to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gadget found"
                    },
                    "404": {
                        "description": "Gadget not found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Gadget"
                ],
                "summary": "Update a gadget",
                "operationId": "88f24f83a2faae6d0438ef130c5ef0b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of gadget to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "example of the body request",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "grapplin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gadget updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Gadget not found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Gadget"
                ],
                "summary": "Delete a gadget",
                "operationId": "40ac3184f1c39ab34b4e0cc0ddef9009",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of gadget to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gadget deleted"
                    },
                    "404": {
                        "description": "Gadget not found"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        },
...
    }
}